+++ src/renderengine/creator/Rect.js
@@ -1,0 +1,24 @@
+import Primitive from './Primitive'
+
+class Rect extends Primitive {
+  getName = 'Rectangle'
+
+  isRect = true
+
+  draw() {
+    const dimensions = this.dimensions.calc()
+
+    if (dimensions.checkMin()) {
+      return
+    }
+
+    this.getColorArray()({
+      posX: dimensions.posX,
+      posY: dimensions.posY,
+      width: dimensions.width,
+      height: dimensions.height,
+    })
+  }
+}
+
+export default Rect
